requirements.txt
coloredlogs
aiobotocore
aiohttp
aiohttp_cors
python-dotenv
PyJWT